remove code related to old notification settings
