+++ static/app/views/settings/account/notifications/notificationSettingsController.tsx
@@ -5,7 +5,6 @@ import type {Organization} from 'sentry/types';
 import withOrganizations from 'sentry/utils/withOrganizations';
 
 import NotificationSettings from './notificationSettings';
-import NotificationSettingsV2 from './notificationSettingsV2';
 
 interface NotificationSettingsControllerProps extends RouteComponentProps<{}, {}> {
   organizations: Organization[];
@@ -13,19 +12,12 @@ interface NotificationSettingsControllerProps extends RouteComponentProps<{}, {}
 }
 
 export function NotificationSettingsController({
-  organizations,
   organizationsLoading,
-  ...props
 }: NotificationSettingsControllerProps) {
   if (organizationsLoading) {
     return <LoadingIndicator />;
   }
-
-  // check if feature is enabled for any organization
-  const hasFeature = organizations.some(org =>
-    org.features.includes('notification-settings-v2')
-  );
-  return hasFeature ? <NotificationSettingsV2 /> : <NotificationSettings {...props} />;
+  return <NotificationSettings />;
 }
 
 export default withOrganizations(NotificationSettingsController);
